Desktop optimizations and nearly done with mobile
Mobile is just about ready, but needs tweaks. Desktop code was adjusted a bit too so it was a bit more organized and would be more responsive at different resolutions.

The navigation currently doesn't work on mobile as the hamburger menu is still being built, but I wanted to get this on here in some form for the meeting.